Fix `Node::getLogCategories()` to work with Dash-specific log categories (#3837)
We use top 32 bits of uint64_t for Dash-specific log categories, see logging.h

+++ src/interfaces/node.cpp
@@ -172,7 +172,7 @@ class NodeImpl : public Node
     void initLogging() override { InitLogging(); }
     void initParameterInteraction() override { InitParameterInteraction(); }
     std::string getWarnings(const std::string& type) override { return GetWarnings(type); }
-    uint32_t getLogCategories() override { return ::logCategories; }
+    uint64_t getLogCategories() override { return ::logCategories; }
     bool baseInitialize() override
     {
         return AppInitBasicSetup() && AppInitParameterInteraction() && AppInitSanityChecks() &&
