+++ src/interfaces/node.h
@@ -123,7 +123,7 @@ public:
     virtual std::string getWarnings(const std::string& type) = 0;
 
     // Get log flags.
-    virtual uint32_t getLogCategories() = 0;
+    virtual uint64_t getLogCategories() = 0;
 
     //! Initialize app dependencies.
     virtual bool baseInitialize() = 0;
